Fixes #179 - InMemoryStateProvider now saves a deep copy
This moves the InMemoryStateProvider to the Proto.Persistence.Tests project as that is the only place it is used.
